Move all game-setup related components to separate folder
